Added Terrain Grid Projector | Refined Unit Selection
Created a grid projector for the terrain (not very visible on the stark
white terrain at the moment, but the terrain should be textured by now).
Completed more of the unit selection system - units of trees can be
individually selected with the mouse, and groups can be selected while
holding shift and holding down the mouse.