Added music, 3D cutting sounds, Pause functionality
Music and ambient sounds are now in the game (finally). Music is set to
be one of two available tracks at random. The sound of the units cutting
is placed in the center of the selected units on the grid, and has a
nice Doppler effect to it.

The pause button works, and is set up so the game can pause everything
without an actual Pause menu coming up (works for message boxes and
being in other menus, just tell the game to pause and it will).

(Other things were added, but I forgot what they were.)